+++ consensus/src/marshal/standard/deferred.rs
@@ -78,11 +78,12 @@ use crate::{
         },
         core::{CommitmentFallback, DigestFallback, Mailbox},
         standard::{
+            relay,
             validation::{
                 await_and_validate_parent, precheck_epoch_and_reproposal, run_app_verify, Decision,
                 ParentCheck,
             },
-            variant, Standard,
+            Standard,
         },
         Update,
     },
@@ -854,7 +855,7 @@ where
     type Plan = Plan<S::PublicKey>;
 
     fn broadcast(&mut self, commitment: Self::Digest, plan: Plan<S::PublicKey>) -> Feedback {
-        variant::broadcast(&self.gates, &self.marshal, commitment, plan)
+        relay::broadcast(&self.gates, &self.marshal, commitment, plan)
     }
 }
