+++ consensus/src/marshal/standard/variant.rs
@@ -6,21 +6,18 @@
 use crate::{
     marshal::{
         ancestry::BlockProvider,
-        application::gates::Gates,
         core::{Buffer, CommitmentFallback, Mailbox, Variant},
     },
-    simplex::{scheme::Scheme as SimplexScheme, Plan},
+    simplex::scheme::Scheme as SimplexScheme,
     types::Round,
     Block,
 };
-use commonware_actor::Feedback;
 use commonware_broadcast::buffered;
 use commonware_codec::Read;
 use commonware_cryptography::{certificate::Scheme, Digestible, PublicKey};
 use commonware_p2p::Recipients;
 use commonware_utils::channel::oneshot;
 use std::{future::Future, sync::Arc};
-use tracing::debug;
 
 /// The standard variant of Marshal, which broadcasts complete blocks.
 ///
@@ -119,36 +116,6 @@ where
 
     fn send(&self, _round: Round, block: Arc<B>, recipients: Recipients<K>) {
         self.broadcast_shared(recipients, block);
-    }
-}
-
-/// Relays a consensus broadcast [`Plan`] through marshal for the standard
-/// variants ([`super::Deferred`] and [`super::Inline`]).
-///
-/// A propose plan sends the staged proposal to all peers, delivering the
-/// durable-sync handle through the staged ack. A forward plan re-sends a
-/// stored block to the requested recipients. A propose plan whose staged
-/// proposal was already consumed falls back to a best-effort forward of the
-/// persisted block.
-pub(super) fn broadcast<S, B>(
-    gates: &Gates<B::Digest, B>,
-    marshal: &Mailbox<S, Standard<B>>,
-    commitment: B::Digest,
-    plan: Plan<S::PublicKey>,
-) -> Feedback
-where
-    S: Scheme,
-    B: Block,
-{
-    match plan {
-        Plan::Propose { round } => {
-            let Some((block, ack)) = gates.take_staged(round, commitment) else {
-                debug!(%round, %commitment, "no staged proposal to relay, attempting forwarding");
-                return marshal.forward(round, commitment, Recipients::All);
-            };
-            marshal.proposed(round, block, Recipients::All, ack)
-        }
-        Plan::Forward { round, recipients } => marshal.forward(round, commitment, recipients),
     }
 }
